refactor: refactoring hash function
