+++ src/cli/selectConsole.js
@@ -5,25 +5,21 @@ export const selectConsole = async (commandLine) => {
   const a = commandLine.trim().split(' ');
   const commandName = a[0];
   switch (commandName) {
-    case 'ls':
-      return await commandList.list(process.cwd());
-    case 'os':
-      commandList.osCommands(a[1]);
-      break;
-    case 'hash':
-      console.log(commandList.calculateHash(a[1]));
-      break;
-    case 'compress':
-      await commandList.compress(a[1], a[2]);
-      break;
-    case  'decompress':
-      await commandList.decompress(a[1], a[2]);
+    case 'up':
+      commandList.myChdir('..');
       break;
     case 'cd':
       commandList.myChdir(a[1]);
       break;
-    case 'up':
-      commandList.myChdir('..');
+    case 'ls':
+      console.log(await commandList.list(process.cwd()));
+      break;
+
+    case 'cat':
+      await commandList.catFile(a[1]);
+      break;
+    case 'add':
+      await commandList.add(a[1]);
       break;
     case 'rn':
       await commandList.rename(a[1], a[2]);
@@ -37,16 +33,26 @@ export const selectConsole = async (commandLine) => {
     case 'rm':
       await commandList.remove(a[1]);
       break;
-    case 'cat':
-      await commandList.catFile(a[1]);
+
+    case 'os':
+      commandList.osCommands(a[1]);
       break;
-    case 'add':
-      await commandList.add(a[1]);
+
+    case 'hash':
+      await console.log(await commandList.calculateHash(a[1]));
+      // await commandList.calcHash2(a[1]);
+      break;
+
+    case 'compress':
+      await commandList.compress(a[1], a[2]);
+      break;
+    case  'decompress':
+      await commandList.decompress(a[1], a[2]);
       break;
 
     default:
       console.log('No command');
-
+      break;
   }
 
   // function promptPrint() {
