+++ src/fs/index.js
@@ -1,6 +1,6 @@
 export { list } from './list.js';
 export { osCommands } from './osCommands.js';
-export { calculateHash } from './calcHash.js';
+export { calculateHash } from './calculateHash.js';
 export { compress } from './compress.js';
 export { decompress } from './decompress.js';
 export { myChdir } from './chdir.js';
