+++ src/fs/osCommands.js
@@ -2,6 +2,8 @@ import * as os from "os";
 
 export const osCommands = (command) => {
 
+  // console.log(command);
+
   const cpuPrint = () => {
     return os.cpus().map((el) => {
       const b =el.model.split('@');
@@ -9,7 +11,8 @@ export const osCommands = (command) => {
     });
   }
 
-  switch (command) {
+  switch (command.trim()) {
+
     case '--EOL':
       console.log(JSON.stringify(os.EOL));
       break;
@@ -20,13 +23,20 @@ export const osCommands = (command) => {
       break;
 
     case '--homedir':
+      console.log(command);
       console.log(`Home directory is: ${os.homedir()}`);
+      break;
+
+    case '--username':
+      console.log(os.userInfo().username);
       break;
 
     case '--architecture':
       console.log(os.arch());
       break;
 
-    default: console.log('Invalid input');
+    default:
+      console.log('Invalid input');
+      break;
   }
 }
